Move remaining oneoffs to osmesa.apps
