+++ src/apps/src/main/scala/osmesa/apps/batch/MergeChangesets.scala
@@ -1,10 +1,13 @@
-package osmesa.analytics.oneoffs
+package osmesa.apps.batch
+
+import java.net.URI
+import java.sql._
+import java.time.Instant
 
 import cats.data.{Validated, ValidatedNel}
 import cats.implicits._
 import com.monovore.decline._
-import io.circe.generic.auto._
-import io.circe.{yaml, _}
+import io.circe._
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
 import org.joda.time.DateTime
@@ -13,19 +16,14 @@ import osmesa.analytics.Analytics
 import vectorpipe.sources.{ChangesetSource, Source}
 import vectorpipe.util.DBUtils
 
-import java.net.URI
-import java.sql._
-import java.time.Instant
-import scalaj.http.Http
-
 /*
  * Usage example:
  *
- * sbt "project analytics" assembly
+ * sbt "project apps" assembly
  *
  * spark-submit \
- *   --class osmesa.analytics.oneoffs.MergeChangesets \
- *   ingest/target/scala-2.11/osmesa-analytics.jar \
+ *   --class osmesa.apps.batch.MergeChangesets \
+ *   ingest/target/scala-2.11/osmesa-apps.jar \
  *   --changesets http://location/of/changeset/replications \
  *   --end-time 1970-01-01T13:00:00Z
  *   s3://path/to/history.orc
@@ -37,8 +35,8 @@ object MergeChangesets
     header = "Bring existing changeset ORC file up to date using changeset stream",
     main = {
 
-      import MergeChangesetUtils._
       import ChangesetSource._
+      import MergeChangesetUtils._
 
       val changesetSourceOpt =
         Opts
