+++ src/apps/src/main/scala/osmesa/apps/streaming/StreamingChangesetMetadataUpdater.scala
@@ -1,4 +1,4 @@
-package osmesa.analytics.oneoffs
+package osmesa.apps.streaming
 
 import java.net.URI
 
@@ -6,20 +6,19 @@ import cats.implicits._
 import com.monovore.decline._
 import org.apache.spark.sql._
 import osmesa.analytics.Analytics
+import osmesa.analytics.stats.ChangesetMetadataForeachWriter
 import vectorpipe.functions._
 import vectorpipe.functions.osm._
 import vectorpipe.sources.Source
-import vectorpipe.util.DBUtils
-import osmesa.analytics.stats.ChangesetMetadataForeachWriter
 
 /*
  * Usage example:
  *
- * sbt "project analytics" assembly
+ * sbt "project apps" assembly
  *
  * spark-submit \
- *   --class osmesa.analytics.oneoffs.StreamingChangesetMetadataUpdater \
- *   ingest/target/scala-2.11/osmesa-analytics.jar \
+ *   --class osmesa.apps.streaming.StreamingChangesetMetadataUpdater \
+ *   ingest/target/scala-2.11/osmesa-apps.jar \
  *   --database-url $DATABASE_URL
  */
 object StreamingChangesetMetadataUpdater
