+++ src/apps/src/main/scala/osmesa/apps/streaming/StreamingUserFootprintTileUpdater.scala
@@ -1,4 +1,4 @@
-package osmesa.analytics.oneoffs
+package osmesa.apps.streaming
 
 import java.io._
 import java.net.URI
@@ -13,11 +13,11 @@ import vectorpipe.sources.Source
 /*
  * Usage example:
  *
- * sbt "project analytics" assembly
+ * sbt "project apps" assembly
  *
  * spark-submit \
- *   --class osmesa.analytics.oneoffs.StreamingUserFootprintUpdater \
- *   ingest/target/scala-2.11/osmesa-analytics.jar
+ *   --class osmesa.apps.streaming.StreamingUserFootprintUpdater \
+ *   ingest/target/scala-2.11/osmesa-apps.jar
  */
 object StreamingUserFootprintTileUpdater
     extends CommandApp(
